requirements.txt
langchain==0.1.12
langchain-openai==0.0.8
langchain-core==0.1.32
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0 